+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 7.2.0
+
+## Features
+- Add method **patchValue** to FormControl, FormGroup and FormArray to update partially the control 
+value.
+
+
 # 7.1.0
 
 ## Features